% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_comp.R
\name{conc_comp}
\alias{conc_comp}
\title{conc_comp Function}
\usage{
conc_comp(conc_name1, conc_name2)
}
\arguments{
\item{conc.name2:}{a character string for the first concentration name as it appears in the 
list of Brown concentrations in the website: 
"https://bulletin.brown.edu/the-college/concentrations/". The input is not case sensitive. 
**conc.name is a character string, so it needs to go with "".}

\item{conc.name2:}{a character string for the second concentration name as it appears in the 
list of Brown concentrations in the website: 
"https://bulletin.brown.edu/the-college/concentrations/". The input is not case sensitive. 
**conc.name is a character string, so it needs to go with "".}
}
\value{
Table of all the required courses for the concentration inputed.
}
\description{
This function takes two concentrations of interest as inputs and returns a 
table with all the required classes for both concentrations.
}
\examples{
conc_rec("Economics")
conc_rec("mathematics")
conc_rec("Math")
conc_rec("Africana Studies")
}
