% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conc_req.R
\name{conc_req}
\alias{conc_req}
\title{conc_rec Function}
\usage{
conc_req(conc_name)
}
\arguments{
\item{conc.name:concentration}{name as it appears in the list of Brown concentrations in the 
website: "https://bulletin.brown.edu/the-college/concentrations/". The input is not case 
sensitive. **conc.name is a character string, so it needs to go with "".}
}
\value{
Table of all the required courses for the concentration inputed.
}
\description{
This function takes the concentration of interest as an input and returns a 
table with all the required classes for that particular concentration.
}
\examples{
conc_rec("Economics")
conc_rec("mathematics")
conc_rec("Africana Studies")
}
