% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sch_req.R
\name{sch_req}
\alias{sch_req}
\title{sch_req Function}
\usage{
sch_req(conc_name, term = c("fall", "spring", "winter"))
}
\arguments{
\item{conc_name:}{Concentration name as it appears in the list of Brown concentrations in the 
website: "https://bulletin.brown.edu/the-college/concentrations/". The input is not case 
sensitive. **conc.name is a character string, so it needs to go with "".}

\item{term:}{The semester of interest: one of "fall", "spring", or "winter".}
}
\value{
Dataframe of the concentration's required course numbers and their titles being offered
in the semester of interest.
}
\description{
This function takes the concentration of interest and a semester as inputs and returns a 
the required classes for the concentration being offered in the specified semester.
}
\examples{
sch_req("Music", "spring")
sch_req("History", "fall")
}
